% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icytracks.R
\docType{package}
\name{icytracks}
\alias{icytracks}
\alias{icytracks-package}
\title{icytracks:  Read Icy Tracks XLS}
\description{
Reads the .xls files generated by the Icy 'Export Tracks to XLS' Track
Processor and convert it to a data.frame.
}

