% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icytracks.R
\name{read.icy.track.xls}
\alias{read.icy.track.xls}
\title{Read an Icy tracks XLS}
\usage{
read.icy.track.xls(file)
}
\arguments{
\item{file}{The Icy XLS file with the tracks.}
}
\description{
Reads the .xls files generated by the Icy 'Export Tracks to XLS' Track
Processor and convert it to a data.frame.
}
\examples{
\dontrun{
read.icy.track.xls('~/icytrack.xls')
}
}

